requirements.txt
openai
pyqt5
python-docx
O365
